handle mongodb indexes and exp policies for upgrades

+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/core/ticket/ProxyGrantingTicketProperties.java
@@ -25,5 +25,5 @@ public class ProxyGrantingTicketProperties implements Serializable {
     /**
      * Maximum length of the proxy granting ticket, when generating one.
      */
-    private int maxLength = 50;
+    private long maxLength = 50;
 }
